Begin adding adapter class for channelworm.PatchClamp
The idea is that an adapter class will provide an interface between the
PyOpenWorm and ChannelWorm data models.
The adapter "has a" PyOpenWorm experiment model, and the ChannelWorm
PatchClamp model "maps" to it through the adapter.

In other words, doing something like `a = adapter(my_cw_model)` gives an
object that has the corresponding PyOpenWorm model. Then if we do
`p = a.pow_model`, "p" is an object in "PyOpenWorm form" that can be
inserted into the PyOW database.

Related: #61